% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bgandilum.R
\docType{data}
\name{bgandilum}
\alias{bgandilum}
\title{Default backgound and illuminant data}
\description{
Default backgound and illuminant data
}
\references{
Endler, J. (1993). The Color of Light in Forests and Its Implications. Ecological Monographs, 63, 1-27.
}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}
}
\keyword{data}
