% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate images}
\usage{
calibrate(image, scale_length = NULL)
}
\arguments{
\item{image}{(required) image data. Either a single image, or a series of images
stored in a list. Preferably the result of \code{\link{getimg}}.}

\item{scale_length}{(required) an integer specifying the length of the scale
in the image.}
}
\value{
an image, or list containing images, for use in further
\code{ptrn} functions, with scales stored as an attribute.
}
\description{
Manually specify the scale of images.
}
\examples{
\dontrun{
# Single image
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'ptrn'))
papilio <- calibrate(papilio, scale = 10)

# Multiple images
snakes <- getimg(system.file("testdata/images/snakes", package = 'ptrn'))
snakes <- calibrate(snakes, scale = 100)
}

}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
