% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_main}
\alias{calibrate_main}
\title{Main function for calibrating images}
\usage{
calibrate_main(image_i, scale_length_i = NULL)
}
\arguments{
\item{image_i}{(required) image data. Either a single image, or a series of images
stored in a list. preferably the result of \code{\link{getimg}}.}

\item{scale_length_i}{(required) an integer specifying the length of the scale
in the image.}
}
\value{
an image, or list containing images, for use in further
\code{pavo} functions, with scales stored as an attribute.
}
\description{
Main function for calibrating images
}
\keyword{internal}
