% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cieplot.R
\name{cieplot}
\alias{cieplot}
\title{CIE plot}
\usage{
cieplot(model, outline = TRUE, col.out = "grey", cex.lab = 1,
  cex.axis = 1, cex.pts = 0.9, ...)
}
\arguments{
\item{model}{(required)}

\item{outline}{Logial. Should the monochromatic loci (the 'horseshoe') be
plotted? Defaults to \code{TRUE}.}

\item{col.out}{Line colour for plotting the monochromatic locus. Defaults
to \code{'grey'}.}

\item{cex.lab}{Character expansion factor for labels.}

\item{cex.axis}{Character expansion factor for minor axis labels.}

\item{cex.pts}{Character expansion factor for stimulus points.}

\item{...}{Additional graphical options. See code{\link{par}}.}
}
\description{
Plot a CIE (XYZ or LAB - todo) chromaticity diagram.
}
\examples{
\dontrun{
data(flowers)
flowers.cie <- cie(flowers, space = 'XYZ', illuminant = 'D65')
cie_plot(flowers.cie)
}
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\references{
Smith T, Guild J. (1932) The CIE colorimetric standards and their use.
   Transactions of the Optical Society, 33(3), 73-134.

Westland S, Ripamonti C, Cheung V. (2012). Computational colour science
   using MATLAB. John Wiley & Sons.
}

