% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_main}
\alias{classify_main}
\title{Main function for identifying colour classes in an image for adjacency analyses}
\usage{
classify_main(imgdat_i, n_cols_i)
}
\arguments{
\item{imgdat_i}{(required) image data. Either a single image, or a series of images
stored in a list. preferably the result of \code{\link{getimg}}.}

\item{n_cols_i}{(required) either an integer, or vector the same length as imgdat (if
passing a list of images), specifying the number of discrete colour classes present
in an image, for k-means clustering.}
}
\value{
A matrix, or list of matrices, of class \code{rimg} containing the colour
class classifications at each pixel location. The RGB values corresponding to
k-means centres (i.e. colour classes) are stored as object attributes.
}
\description{
Main function for identifying colour classes in an image for adjacency analyses
}
\keyword{internal}
