% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rimg.R
\name{classplot}
\alias{classplot}
\title{Plot colour-classified images}
\usage{
classplot(x, ...)
}
\arguments{
\item{x}{(required) colour-classified image data, or a list thereof. Preferably
the result of \code{\link{classify}}.}

\item{...}{additional graphical parameters passed to \code{\link{image}}.}
}
\description{
Plot colour-classified images
}
\keyword{internal}
