% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipspec.R
\name{clipspec}
\alias{clipspec}
\title{Clip artifacts from spectra}
\usage{
clipspec(rspecdata, from, to, interpolate = FALSE)
}
\arguments{
\item{rspecdata}{(required) a data frame, possibly of class \code{rspec}, which
contains a column containing a wavelength range, named 'wl', and spectra
data in remaining columns.}

\item{from, to}{range of wavelengths to clip out. This range is exclusive,
meaning that wavelengths equal to \code{from} or \code{to} will not be clipped. \code{to}
must be greater or equal to \code{from}.}

\item{interpolate}{(logical) whether to linearly interpolate reflectance
values across missing wavelengths once the latter have been removed. If
FALSE (default), missing wavelengths are not replaced.}
}
\value{
A data frame of class \code{rspec} with the clipped data.
}
\description{
Cuts out portions of spectra that may have been identified as containing
artifactual reflectance values.
}
\details{
The interpolation step internally calls the \code{as.rspec()} function
with its \code{interp} argument set to TRUE.
}
\note{
Preferably use this function before \code{procspec()}, otherwise artifacts
may affect the smoothing.
}
\examples{

# Load data
data(sicalis)

# Eyeball
plot(sicalis, select = 10:14)

# Remove the bump around 470nm
sicalis <- clipspec(sicalis, from = 460, to = 480)

# Check again
plot(sicalis, select = 10:14)

}
\seealso{
\code{\link[=procspec]{procspec()}}, \code{\link[=as.rspec]{as.rspec()}}
}
\author{
Raphaël Scherrer (\email{raphael.scherrer@evobio.eu})
}
