% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cocplot.R
\name{cocplot}
\alias{cocplot}
\title{Plot the colour opponent coding diagram}
\usage{
cocplot(cocdata, labels = TRUE, cex.labels = 0.9, tick.loc = c(-12, -9,
  -6, -3, 3, 6, 9, 12), ...)
}
\arguments{
\item{cocdata}{(required) a data frame, possibly a result from the \code{categorical}
function, containing values for 'x' and 'y' coordinates as columns (labeled as such)}

\item{labels}{plot axis labels? Defaults to \code{TRUE}}

\item{cex.labels}{character expansion factor for category labels when \code{labels = TRUE})}

\item{tick.loc}{a numeric vector specifying the location of tick marks on x & y axes}

\item{...}{additional graphical options. See code{\link{par}}}
}
\description{
\code{cocplot} produces a plot based on the colour opponent coding diagram of Backhaus (1991)
}
\examples{
\dontrun{
data(flowers)
vis.flowers <- vismodel(flowers, visual = 'apis', qcatch = 'Ei', relative = FALSE, vonkries = TRUE)
coc.flowers <- coc(vis.flowers)
cocplot(coc.flowers)
}
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\references{
Backhaus W. (1991). Color opponent coding in the visual system
 of the honeybee. Vision Research, 31, 1381-1397.
}

