% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diplot.R
\name{diplot}
\alias{diplot}
\title{Plot a dichromat segment}
\usage{
\code{plot(clrspdata)}.
}
\arguments{
\item{didata}{(required) a data frame, possibly a result from the \code{dispace} 
function, containing values for the 'x' coordinates as a column (labeled as such)}

\item{labels}{plot verticy labels? Defaults to \code{TRUE}}

\item{achro}{should a point be plotted at the origin (defaults to \code{TRUE})?}

\item{achrocol}{color of the point at the origin \code{achro = TRUE} (defaults to grey)}

\item{achrosize}{size of the point at the origin when \code{achro = TRUE} (defaults to 0.8)}

\item{cex.labels}{character expansion factor for category labels when \code{labels = TRUE})}

\item{out.lwd}{line width for segment (defaults to 1)}

\item{out.lcol}{line colour for segment (defaults to black)}

\item{out.lty}{line type for segment (defaults to 1)}

\item{...}{additional graphical options. See \code{\link{par}}.}
}
\description{
\code{diplot} produces a dichromat segment plot.
}
\examples{
\dontrun{
data(flowers)
vis.flowers <- vismodel(flowers, visual = 'canis')
di.flowers <- colspace(vis.flowers, space = 'di')
plot(di.flowers)
}

}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\references{
Kelber A, Vorobyev M, Osorio D. (2003). Animal colour vision
   - behavioural tests and physiological concepts. Biological Reviews, 78,
   81 - 118.
}
\keyword{internal}

