% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-flowers.R
\docType{data}
\name{flowers}
\alias{flowers}
\title{Reflectance spectra from a suite of native Australian flowers, collected
around Cairns, Queensland.}
\format{An object of class \code{rspec} (inherits from \code{data.frame}) with 401 rows and 37 columns.}
\usage{
flowers
}
\description{
Dataset containing reflectance measurements from 36 native Australian
angiosperm species, indicated by column names.
}
\references{
Dalrymple, R., L., Kemp, D. J., Flores-Moreno, H., Laffan, S. W.,
White, T. E., Hemmings, F. A., Tindall, M. L., & Moles, A. T. (2015). Birds,
butterflies and flowers in the tropics are not more colourful than those at higher
latitudes. Global Ecology and Biogeography, 24(12), 1424-1432. \doi{10.1111/geb.12368}

White, T. E., Dalrymple, R. L., Herberstein, M. E., & Kemp, D. J. (2017).
The perceptual similarity of orb-spider prey lures and flowers colours. Evolutionary
Ecology, 31(1), 1-20. \doi{10.1007/s10682-016-9876-x}

Dalrymple, R. L., Flores-Moreno, H., Kemp, D. J., White, T. E.,
Laffan, S. W., Hemmings, F. A., Tindall, M. L., & Moles, A. T. (2018). Abiotic
and biotic predictors of macroecological patterns in bird and butterfly coloration.
Ecological Monographs, 88(2), 204-224. \doi{10.1002/ecm.1287}
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\keyword{datasets}
