% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getimg.R
\name{getimg}
\alias{getimg}
\title{Import image data}
\usage{
getimg(imgpath = getwd(), subdir = FALSE)
}
\arguments{
\item{imgpath}{(required) either the full path to a given image (including extension),
or the path to a folder in which multiple image files are located. Mixed file formats
within a folder are fine.}

\item{subdir}{should subdirectories within the \code{where} folder be
included in the search? (defaults to \code{FALSE}).}
}
\value{
a array, or list of arrays, of class \code{rimg}, containing image data 
for use in further \code{pavo} functions.
}
\description{
Finds and imports PNG, JPEG, or BMP images.
}
\examples{
\dontrun{
# Single image
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))

# Multiple images
snakes <- getimg(system.file("testdata/images/snakes", package = 'pavo'))
}

}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
