% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.colspace.R
\name{is.colspace}
\alias{is.colspace}
\title{Test if object is of class 'colspace'}
\usage{
is.colspace(object)
}
\arguments{
\item{object}{an R object}
}
\value{
a logical value indicating whether the object is of class \code{colspace}
}
\description{
Test if object is of class 'colspace'
}
\seealso{
\code{\link[=colspace]{colspace()}}
}
