% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rimg.r
\name{is.rimg}
\alias{is.rimg}
\title{Check if data is an rimg object.}
\usage{
is.rimg(object)
}
\arguments{
\item{object}{(required) a three-dimensional array containing RGB values.}
}
\value{
a logical value indicating whether the object is of class
\code{rimg}
}
\description{
Check if data is an rimg object.
}
