% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.vismodel.R
\name{is.vismodel}
\alias{is.vismodel}
\title{Test if object is of class 'vismodel'}
\usage{
is.vismodel(object)
}
\arguments{
\item{object}{an \R object}
}
\value{
a logical value indicating whether the object is of class \code{vismodel}.
}
\description{
Test if object is of class 'vismodel'
}
\seealso{
\code{\link[=vismodel]{vismodel()}}
}
