% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendtetra.R
\name{legendtetra}
\alias{legendtetra}
\title{Add Legend to a Static Tetrahedral Colorspace
\code{legendtetra} adds legend to a 
static tetrahedral colorspace plot.}
\usage{
legendtetra(location = c(-1.4, -0.3, 0.7), ...)
}
\arguments{
\item{location}{a vector containing x, y and z values for the position of
the legend.}

\item{...}{additional arguments passed to \code{\link{legend}}.}
}
\value{
\code{legendtetra} adds a legend to astatic tetrahedral colorspace plot. 
for additional information on which areguments are necessary and how they are used,
see \code{\link{legend}}.
}
\description{
Add Legend to a Static Tetrahedral Colorspace
\code{legendtetra} adds legend to a 
static tetrahedral colorspace plot.
}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}
}

