% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lines.rspec}
\alias{lines.rspec}
\title{Add lines to an existing spectrum plot}
\usage{
\method{lines}{rspec}(x, select = NULL, type = "overlay", ...)
}
\arguments{
\item{x}{(required) a data frame, possibly an object of class \code{rspec},
with a column with wavelength data, named `wl', and the remaining column containing
spectra to plot.}

\item{select}{specification of which spectra to plot. Can be a numeric vector or
factor (e.g., \code{sex=='male'})}

\item{...}{additional arguments passed to lines}
}
\description{
Connecting reflectance spectra datapoints with line segments.
}
\examples{
\dontrun{
data(teal)
plot(teal, select = 2:4, col = 'red')
lines(teal, select = 5:8, col = 'blue')
}
}
\author{
Chad Eliason \email{cme16@zips.uakron.edu}
}
\seealso{
\code{\link{plot}}, \code{\link{lines}}
}

