% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nspace.R
\name{nspace}
\alias{nspace}
\title{n-chromatic colour space}
\usage{
nspace(vismodeldata)
}
\arguments{
\item{vismodeldata}{(required) quantum catch color data. Can be either the
result from \code{\link[=vismodel]{vismodel()}} or independently calculated data}
}
\description{
Calculates coordinates and colourimetric variables that represent reflectance
spectra in a n-chromatic colour space.
}
\examples{
fakemantisshrimp <- sensmodel(c(325, 350, 400, 425, 450, 500, 550, 600, 650, 700), beta = FALSE, integrate = FALSE)

data(flowers)
vis_flowers <- vismodel(flowers, visual = fakemantisshrimp)
colsp_flowers <- nspace(vis_flowers)
}
\references{
Pike, T.W. Generalised Chromaticity Diagrams for Animals with n-chromatic
Colour Vision. J Insect Behav 25, 277–286 (2012).
\doi{10.1007/s10905-011-9296-2}
}
