% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procimg.R
\name{outliner}
\alias{outliner}
\title{Internal function for selecting focal-stimulus outline}
\usage{
outliner(image_i, smooth_i, iterations_i, ...)
}
\arguments{
\item{image_i}{(required) image data. Either a single image, or a series of images
stored in a list. preferably the result of \code{\link{getimg}}.}

\item{smooth_i}{should the polygon specified when \code{outline = TRUE} be smoothed
using Chaikin's corner-cuting algorithm? Defaults to \code{FALSE}.}

\item{iterations_i}{the number of smoothing iterations, when \code{smooth = TRUE}.}

\item{...}{additional graphical parameters. Also see \code{\link{par}}.}
}
\value{
an image, or list containing images, for use in further
\code{pavo} functions, with scales stored as an attribute.
}
\description{
Internal function for selecting focal-stimulus outline
}
\keyword{internal}
