% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_avantes.R
\name{parse_avantes}
\alias{parse_avantes}
\title{Import Avantes binary file}
\usage{
parse_avantes(filename)
}
\arguments{
\item{path}{Path of the ABS, TRM or ROH file}
}
\value{
A dataframe with 2 columns: first column contains the wavelengths and
second column contains the processed spectrum
}
\description{
Internal function used by \code{\link{getspec}} to parse Avantes binary files
generated by the Avasoft software.
}
\seealso{
\code{\link{getspec}}
}
\author{
Hugo Gruson \email{hugo.gruson+R@normalesup.org}
}
\keyword{internal}
