% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_procspec.R
\name{parse_procspec}
\alias{parse_procspec}
\title{Import ProcSpec spectra file}
\usage{
parse_procspec(path)
}
\arguments{
\item{path}{Path of the ProcSpec file}
}
\value{
A dataframe with 2 columns: first column contains the wavelengths and
second column contains the processed spectrum
}
\description{
Internal function used by \code{\link{getspec}} to parse ProcSpec files
generated by the OceanOptics software spectrasuite.
}
\seealso{
\code{\link{getspec}}
}
\author{
Hugo Gruson \email{hugo.gruson+R@normalesup.org}
}
