% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pavo-deprecated.R
\docType{package}
\name{pavo-deprecated}
\alias{pavo-deprecated}
\alias{pavo-deprecated-package}
\alias{tcs}
\title{Deprecated function(s) in the pavo package}
\usage{
tcs(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the pavo package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
  \code{colspace} \tab now a synonym for \code{\link{tcs}}\cr
}
}

