% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pavo-package.R
\docType{package}
\name{pavo-package}
\alias{pavo}
\alias{pavo-package}
\title{pavo: Perceptual Analysis, Visualization and Organization of Spectral Colour Data}
\description{
A cohesive framework for the spectral and spatial analysis of colour described in Maia, Eliason, Bitton, Doucet & Shawkey (2013) \doi{10.1111/2041-210X.12069} and Maia, Gruson, Endler & White (2019) \doi{10.1111/2041-210X.13174}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pavo.colrverse.com}
  \item \url{https://github.com/rmaia/pavo/}
  \item Report bugs at \url{https://github.com/rmaia/pavo/issues}
}

}
\author{
\strong{Maintainer}: Thomas White \email{thomas.white026@gmail.com} (\href{https://orcid.org/0000-0002-3976-1734}{ORCID})

Authors:
\itemize{
  \item Thomas White \email{thomas.white026@gmail.com} (\href{https://orcid.org/0000-0002-3976-1734}{ORCID})
  \item Rafael Maia \email{rafa.maia@gmail.com} (\href{https://orcid.org/0000-0002-7563-9795}{ORCID})
  \item Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})
  \item John Endler \email{john.endler@deakin.edu.au}
  \item Chad Eliason \email{cme16@zips.uakron.edu}
  \item Pierre-Paul Bitton \email{bittonp@uwindsor.ca} (\href{https://orcid.org/0000-0001-5984-2331}{ORCID})
}

}
\keyword{internal}
