% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.colorspace.r
\name{plot.colorspace}
\alias{plot.colorspace}
\title{Plot spectra in various colourspaces}
\usage{
usage
}
\arguments{
\item{...}{additional arguments passed to plot.}

\item{x}{(required) a data frame, possibly an object of class \code{colorspace},
blah}

\item{foo}{\itemize{
 \item \code{stuff}
}}

\item{sectors}{plot the bee-hue sector dividers? Options are:
   \itemize{
       \item \code{'none'}: No sectors (default)
       \item \code{'fine'}: 36 10-degree sectors
       \item \code{'coarse'}: six bee-hue sectors (blue, blue-green, green, uv-green, uv, uv-blue).
       }}

\item{Hoopah}{}
}
\description{
yup
}
\details{
details yo
}
\note{
Some obly available blah:
}
\section{Things section}{

}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{plot}}
}

