% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rimg.R
\name{plot.rimg}
\alias{plot.rimg}
\title{Plot unprocessed or colour-classified images}
\usage{
\method{plot}{rimg}(x, ...)
}
\arguments{
\item{x}{(required) unprocessed or colour-classified image data, or a list thereof. 
Preferably the result of \code{\link{classify}}.}

\item{...}{additional graphical parameters. See \code{\link{rawplot}} for unprocessed
image plotting, and \code{\link{classplot}} for colour-classified image plots.
Also see \code{\link{par}}.}
}
\value{
a image plot or plots.
}
\description{
Plot unprocessed or colour-classified image data. If the
input data is a list of classified image data, images will be stepped through
one by one.
}
\examples{
\dontrun{
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))
plot(papilio)
papilio_class <- classify(papilio, n_cols = 4)
plot(papilio_class)

# Multiple images
snakes <- getimg(system.file("testdata/images/snakes", package = 'pavo'))
snakes_class <- classify(snakes, n_cols = 3)
plot(snakes_class)
}

}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
