% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vismodel.R
\name{plot.vismodel}
\alias{plot.vismodel}
\title{Title}
\usage{
\method{plot}{vismodel}(vismodeldata, colour.solid = TRUE, ...)
}
\arguments{
\item{vismodeldata}{the result from \code{\link[=vismodel]{vismodel()}}. Quantum catches computed
with another tool are not yet supported.}

\item{colour.solid}{logical (defaults to \code{TRUE}). Should the colour solid for
the relevant illuminant and visual system be plotted.}

\item{...}{additional arguments passed to \code{\link[=plot]{plot()}}}
}
\value{

}
\description{
Title
}
\details{
This function only works with non-relative quantum catches
(\code{relative = FALSE} in \code{\link[=vismodel]{vismodel()}}) as it would always be the line y = 1 -x
otherwise.
}
\examples{
data(sicalis)
vis_sicalis <- vismodel(sicalis, "canis", relative = FALSE)
plot(vis_sicalis)

}
\references{
Wilkins, L., & Osorio D. C. (2019). Object colours, material
properties and animal signals. Journal of Experimental Biologyn 222(21),
\doi{10.1242/jeb.204487}.
}
