% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent.R
\name{polymask}
\alias{polymask}
\title{Manipulate classified image data that fall inside/outside a polygon}
\usage{
polymask(imagedat, poly, alterWhich = c("inside", "outside"))
}
\arguments{
\item{imagedat}{data.}

\item{poly}{xy polygon coordinates.}

\item{alterWhich}{manipulate values inside or outside the polygon.}
}
\description{
Manipulate classified image data that fall inside/outside a polygon
}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
\keyword{internal}
