% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rimg.R
\name{rawplot}
\alias{rawplot}
\title{Plot unprocessed images}
\usage{
rawplot(x, ...)
}
\arguments{
\item{x}{(required) colour-classified image data, or a list thereof. Preferably
the result of \code{\link{classify}}.}

\item{...}{additional graphical parameters plassed to \code{\link{plot}}.}
}
\description{
Plot unprocessed images
}
\keyword{internal}
