% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sicalis.R
\docType{data}
\name{sicalis}
\alias{sicalis}
\title{Spectral curves from three body regions of stripe-tailed yellow finch
(\emph{Sicalis citrina}) males}
\format{
An object of class \code{rspec} (inherits from \code{data.frame}) with 401 rows and 22 columns.
}
\usage{
data(sicalis)
}
\description{
Dataset containing reflectance measurements from 3 body parts ("C": crown,
"B": breast, "T": throat) from seven male stripe-tailed yellow finches
(\emph{Sicalis citrina})
}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}
}
\keyword{datasets}
