% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spec2.R
\name{simulate_spec2}
\alias{simulate_spec2}
\title{Simulate a reflectance spectrum}
\usage{
simulate_spec2(
  wl.inflect = NULL,
  wl.peak = NULL,
  width.sig = 20,
  width.gauss = 70,
  xlim = c(300, 700),
  ylim = c(0, 100)
)
}
\arguments{
\item{wl.inflect}{A numeric vector specifying the wavelength location (in nm)
for one or more inflection point(s) for a 'sigmoid' shaped curve, if desired.}

\item{wl.peak}{A numeric vector specifying the wavelength location (in nm)
for one or more inflection point(s) for a 'Gaussian' (or 'bell') shaped curve,
if desired.}

\item{width.sig}{A numeric value or vector (if multiple wl.inflect values are specified)
specifying the steepness of the change, for any sigmoidal curve(s). Required when
\code{wl-peak} is specified. Defaults to 20 nm.}

\item{width.gauss}{A numeric value or vector specifying the the full-width at half-maximum
of any Gaussian curve(s). Required when \code{wl-peak} is specified. Defaults to 70 nm.}

\item{xlim}{A vector specifying the wavelength range of the simulated spectra. Defaults to
300-700nm (\code{c(300, 700)}).}

\item{ylim}{A vector specifying the minimum and maximum reflectance values of the resulting
curve. Defaults to 0 - 100 \% (\code{c(0, 100)}).}
}
\value{
An object of class \code{rspec}.
}
\description{
Simulate a natural reflectance spectrum with sigmoidal (s-shaped) and/or Gaussian
(bell-shaped) curves. Multiple Gaussian and sigmoidal curves can be combined in
a single spectrum, to simulate more complex reflectance functions.
}
\examples{
# Single sigmoidal spectrum, with an inflection at 500 nm.
spec <- as.rspec(simulate_spec2(wl.inflect = 500))

# Single Gaussian spectrum, with a peak at 400 nm
spec2 <- as.rspec(simulate_spec2(wl.peak = 400))

# Combination of both Gaussian (with peak at 340 nm) and sigmoidal (with inflection
# at 560 nm)
spec3 <- as.rspec(simulate_spec2(wl.peak = 340, wl.inflect = 560))

# Double-Gaussian peaks of differing widths
spec4 <- as.rspec(simulate_spec2(wl.peak = c(340, 560), width.gauss = c(12, 40)))

# Complex spectrim with multi-Gaussian and single sigmoidal peak
spec5 <- as.rspec(simulate_spec2(wl.peak = c(340, 430), width.gauss = c(20, 60), wl.inflect = 575)) 

}
\author{
Thomas White \email{thomas.white026@gmail.com}

Hugo Gruson \email{hugo.gruson+R@normalesup.org}
}
