% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rimg.R
\name{summary.rimg}
\alias{summary.rimg}
\title{Colour-classified image summary}
\usage{
\method{summary}{rimg}(object, plot = FALSE, ...)
}
\arguments{
\item{object}{(required) Results of \code{\link{classify}}.}

\item{plot}{logical; plot the actual colours corresponding to colour class
categories? Defaults to \code{FALSE}.}

\item{...}{class consistency (ignored)}
}
\value{
Either the RGB values of the k-means centres from the colour-classified image,
or a plot of the colours specified by those values (when \code{plot = TRUE}.
}
\description{
Returns the attributes of the colour-classified matrix generated from \code{\link{classify}}.
}
\examples{
\dontrun{
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))
papilio_class <- classify(papilio, n_cols = 4)
summary(papilio_class)

# Plot the unprocessed and colour-classified images alongside the colour class samples
par(mfrow = c(1,3))
plot(papilio)
plot(papilio_class)
summary(papilio_class, plot = TRUE)
}

}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
