% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcsshape.R
\name{tcsshape}
\alias{tcsshape}
\title{Produces a 3D alphashape in tetrahedral colour space}
\usage{
tcsshape(
  tcsdata,
  col = "black",
  alpha = 0.2,
  grid.alpha = 1,
  grid = TRUE,
  fill = TRUE,
  lwd = 1,
  avalue
)
}
\arguments{
\item{tcsdata}{(required) a data frame, possibly a result from the
\code{\link[=colspace]{colspace()}} or \code{\link[=tcspace]{tcspace()}} function, containing values for the 'x', 'y'
and 'z' coordinates as columns (labeled as such).}

\item{col}{colour of the points in the plot (defaults to black)}

\item{alpha}{transparency of points (or volume fill in \code{\link[=tcsvol]{tcsvol()}})}

\item{grid.alpha}{transparency of the volume polygon grid lines}

\item{grid}{if \code{TRUE}, connects the polygon outlining the volume occupied by points (defaults to \code{TRUE})}

\item{fill}{if \code{TRUE}, fills the volume occupied by points (WARNING: transparency
is not saved properly if exported using \code{rgl.postscript})(defaults to \code{TRUE}).}

\item{lwd}{graphical parameters for the edges of the tetrahedron.}

\item{avalue}{Value of alpha parameter for the alphashape}
}
\description{
Produces a 3D alphashape in tetrahedral colour space
}
\examples{
\dontrun{
# For plotting
data(sicalis)
vis.sicalis <- vismodel(sicalis, visual = "avg.uv")
tcs.sicalis <- colspace(vis.sicalis, space = "tcs")
tcsplot(tcs.sicalis, col = "blue", size = 0.005)
tcsshape(tcs.sicalis, avalue = 0.1)
}

}
