% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-teal.R
\docType{data}
\name{teal}
\alias{teal}
\title{Angle-resolved reflectance data for the iridescent wing patch of a male green-winged
teal (\emph{Anas carolinensis})}
\format{An object of class \code{rspec} (inherits from \code{data.frame}) with 401 rows and 13 columns.}
\usage{
data(teal)
}
\description{
Dataset containing reflectance measurements from the wing patch of a single
male at different incident angles (15-75 degrees in 5-degree increments).
}
\author{
Chad Eliason \email{cme16@zips.uakron.edu}
}
\keyword{datasets}
