% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetraplot.R
\name{tetraplot}
\alias{tetraplot}
\title{Plot a Static Tetrahedral Colorspace}
\usage{
tetraplot(tcsdata, vertexsize = 0.8, achro = TRUE, achrosize = 0.8,
  achrocol = "grey", linwd = 1, lincol = "darkgrey", view = 70, ...)
}
\arguments{
\item{tcsdata}{(required) a data frame, possibly a result from the \code{colspace} 
function, containing values for the 'x', 'y' and 'z' coordinates as columns (labeled as such)}

\item{vertexsize}{size of the points at the vertices}

\item{achrosize}{size of the point in the achromatic center}

\item{achrocol}{color of the point in the achromatic center}

\item{linwd}{line width for the edges of the tetrahedron}

\item{lincol}{line color for the edges of the tetrahedron}

\item{view}{orientation of the tetrahedron in degrees (defaults to 70)}
}
\value{
\code{tcsplot} creates a 3D plot using functions of the package \code{scatterplot3d}.
}
\description{
\code{tcsplot} produces a static 3D plot of a tetrahedral 
 colorspace. Accessed via the function \code{\link{plot.colspace}}.
}
\examples{
\dontrun{

# For plotting
data(sicalis)
vis.sicalis <- vismodel(sicalis, visual = 'avg.uv')
tcs.sicalis <- colspace(vis.sicalis, space = 'tcs')
plot(tcs.sicalis)

}

}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}

Thomas White \email{thomas.white026@gmail.com}
}
\references{
Stoddard, M. C., & Prum, R. O. (2008). Evolution of avian plumage 
 color in a tetrahedral color space: A phylogenetic analysis of new world buntings. 
 The American Naturalist, 171(6), 755-776.

Endler, J. A., & Mielke, P. (2005). Comparing entire colour patterns 
 as birds see them. Biological Journal Of The Linnean Society, 86(4), 405-431.
}
\seealso{
\code{\link[rgl]{spheres3d}},\code{\link[rgl]{rgl.postscript}}, 
\code{\link[rgl]{rgl.snapshot}},\code{\link[rgl]{rgl.material}}
}
\keyword{internal}

