% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrashape.R
\name{tetrashape}
\alias{tetrashape}
\title{Plot a (possibly concave) tetrahedral colour space}
\usage{
tetrashape(
  tcsdata,
  alpha = 0.2,
  grid = TRUE,
  fill = TRUE,
  new = FALSE,
  avalue,
  ...
)
}
\arguments{
\item{tcsdata}{(required) a data frame, possibly a result from the
\code{\link[=colspace]{colspace()}} or \code{\link[=tcspace]{tcspace()}} function, containing values for the 'x', 'y'
and 'z' coordinates as columns (labeled as such).}

\item{alpha}{transparency of volume (if \code{fill = TRUE}).}

\item{grid}{logical. if \code{TRUE} (default), draws the polygon outline defined by the points.}

\item{fill}{logical. if \code{TRUE} (default), fills the volume defined by the points.}

\item{new}{logical. Should a new plot be started or draw over an open plot?
(defaults to \code{FALSE})}

\item{avalue}{Value of alpha parameter for the alphashape}

\item{...}{additional graphical options. See \code{\link[=polygon]{polygon()}} and \code{\link[=tetraplot]{tetraplot()}}.}
}
\description{
Produces a 3D alphashape in tetrahedral colour space when plotting a
non-interactive tetrahedral plot.
}
\examples{


# For plotting
data(sicalis)
vis.sicalis <- vismodel(sicalis, visual = "avg.uv")
tcs.sicalis <- colspace(vis.sicalis, space = "tcs")
plot(tcs.sicalis)
tetrashape(tcs.sicalis, avalue = 0.1)

}
