% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transmissiondata.R
\docType{data}
\name{transmissiondata}
\alias{transmissiondata}
\title{Default ocular transmission data}
\description{
Default ocular transmission data
}
\references{
Hart, N. S., Partridge, J. C., Cuthill, I. C., Bennett, A. T. D. (2000). 
Visual pigments, oil droplets, ocular media and cone photoreceptor distribution in two
species of passerine
}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}
}
\keyword{data}
