% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol.colspace.r
\name{vol.colspace}
\alias{vol.colspace}
\title{Plot a 3D Convex Hull in a Tetrahedral Colorspace}
\usage{
vol.colspace(clrspdata, ...)
}
\value{
\code{vol.colspace} creates a 3D convex hull within a \code{plot.colspace} object, 
 only when \strong{\code{space = 'tcs'}}.
}
\description{
\code{vol.colspace} produces a 3D convex hull in tetrahedral color space (i.e. when 
\code{space = 'tcs'})
}

