% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent.R
\name{adjacent}
\alias{adjacent}
\title{Run an adjacency/boundary strength analysis}
\usage{
adjacent(classimg, xscale = NULL, xpts = 100, bkgID = NULL,
  polygon = NULL, exclude = c("none", "background", "object"),
  coldists = NULL, hsl = NULL, cores = getOption("mc.cores", 2L))
}
\arguments{
\item{classimg}{(required) an xyz image matrix, or list of matrices, in which
x and y correspond to pixel coordinates, and z is a numeric code specifying
a colour-class. Preferably the result of \code{\link{classify}}.}

\item{xscale}{(required) an integer specifying the true length of the x-axis,
in preferred units. Not required, and ignored, only if image scales have been set via
\code{\link{procimg}}.}

\item{xpts}{an integer specifying the number of sample points along the x and y axes,
from which the square sampling grid is constructed. Defaults to 100 (i.e for a 100
x 100 grid), though this should be carefully considered.}

\item{bkgID}{an integer or vector specifying the colour-class ID number(s) of
pertaining to the background alone, for relatively homogeneous and uniquely-identified
backgrounds (e.g. the matte background of pinned specimens). Examine the attributes of, or
call \code{summary} on, the result of \code{\link{classify}} to visualise the RGB
values corresponding to colour-class ID numbers. Ignored if the focal object
and background has been identified using \code{\link{procimg}}.}

\item{polygon}{a data.frame of x-y coordinates delineating a closed polygon that
separates the focal object from the background. Not required, and ignored, if the
focal object outline is specified using \code{\link{procimg}}.}

\item{exclude}{The portion of the image to be excluded from the analysis, if any.
If excluding the focal object, its outline must first have been idenfitied using
\code{\link{procimg}}. If excluding the image background it must either have been
identified using \code{\link{procspec}}, or if it is relatively homogeneous,
the colour-class ID's uniquely corresponding to the background can be specified using
\code{bkgID}.}

\item{coldists}{A data.frame specifying the visually-modelled chromatic (dS)
and/or achromatic (dL) distances between colour-categories. The first two columns
should be named 'c1' and 'c2', and specify all possible combinations of colour
category ID's (NOTE: default ID's follow the convention 'clr1, 'clr2', ...),
with the remaining columns named dS (for chromatic distances) and/or dL
(for achromatic distances). See \code{\link{vismodel}} and \code{\link{colspace}}
for visual modelling with spectral data.}

\item{hsl}{data.frame specifying the hue, saturation, and luminance of color patch elements,
as might be estimated via \code{\link{vismodel}} and \code{\link{colspace}}. The first
column, named 'patch', should contain color category ID's (NOTE: default ID's follow the convention
'clr1, 'clr2', ...), with the remain columns specifying one or more of 'hue' (angle, in radians),
'sat', and/or 'lum'.}

\item{cores}{number of cores to be used in parallel processing. If \code{1}, parallel
computing will not be used. Defaults to \code{getOption("mc.cores", 2L)}.}
}
\value{
a data frame of summary variables:
\itemize{
  \item \code{'k'}: The number of user-specified colour and/or luminance classes.
  \item \code{'N'}: The grand total (sum of diagonal and off-diagonal) transitions.
  \item \code{'n_off'}: The total off-diagonal transitions.
  \item \code{'p_i'}: The overall frequency of colour class \emph{i}.
  \item \code{'q_i_j'}: The frequency of transitions between \emph{all} colour classes
   \emph{i} and \emph{j}, such that \code{sum(q_i_j) = 1}.
  \item \code{'t_i_j'}: The frequency of off-diagonal (i.e. class-change
    transitions) transitions \emph{i} and \emph{j}, such that \code{sum(t_i_j) = 1}.
  \item \code{'m'}: The overall transition density (mean transitions),
    in units specified in the argument \code{xscale}.
  \item \code{'m_r'}: The row-wise transition density (mean row transitions),
    in user-specified units.
  \item \code{'m_c'}: The column-wise transition density (mean column transitions),
    in user-specified units.
  \item \code{'A'}: The transition aspect ratio (< 1 = wide, > 1 = tall).
  \item \code{'B'}: The animal/background transition ratio.
  \item \code{'Sc'}: Simpson colour class diversity.
  \item \code{'St'}: Simpson transition diversity.
  \item \code{'Jc'}: Simpson colour class diversity relative to its achievable maximum.
  \item \code{'Jt'}: Simpson transition diversity relative to its achievable maximum.
  \item \code{'Rt'}: Ratio of animal-animal and animal-background transition diversities,
  \code{Rt = St_a_a / St_a_b}.
  \item \code{'Rab'}: Ratio of animal-animal and background-background transition diversities,
  \code{Rt = St_a_a / St_b_b}.
  \item \code{'m_dS', 's_dS', 'cv_dS'}: weighted mean, sd, and coefficient of variation of the chromatic boundary strength.
  \item \code{'m_dL', 's_dL', 'cv_dL'}: weighted mean, sd, and coefficient of variation of the achromatic boundary strength.
  \item \code{'m_hue', 's_hue', 'var_hue'}: circular mean, sd, and variance of overall pattern hue (in radians).
  \item \code{'m_sat', 's_sat', 'cv_sat'}: weighted mean, sd, and coefficient variation of overall pattern saturation.
  \item \code{'m_lum', 's_lum', 'cv_lum'}: weighted mean, sd, and coefficient variation of overall pattern luminance.
  }
}
\description{
Calculate summary variables from the adjacency (Endler 2012) and
boundary-strength (Endler et al. 2018) and overall-pattern contrast
(Endler & Mielke 2005) analyses.
}
\examples{
\dontrun{
# Single image
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))
papilio_class <- classify(papilio, kcols = 4)
papilio_adj <- adjacent(papilio_class, xpts = 150, xscale = 100)

# Single image, with fake color distances and hsl values
# Fake color distances
distances <- data.frame(c1 = c('clr1', 'clr1', 'clr1', 'clr2', 'clr2', 'clr3'),
                        c2 = c('clr2', 'clr3', 'clr4', 'clr3', 'clr4', 'clr4'),
                        dS = c(5, 3, 5, 2, 6, 3),
                        dL = c(5.5, 6.6, 3.3, 2.2, 4.4, 6.6))

# Fake hue, saturation, luminance values
hsl_vals <- data.frame(patch = c('clr1', 'clr2', 'clr3', 'clr4'),
                       hue = c(1.5, 2.2, 1.0, 0.5),
                       lum = c(10, 5, 7, 3),
                       sat = c(3.5, 1.1, 6.3, 1.3))

# Full analysis
papilio_adj <- adjacent(papilio_class, xpts = 150, xscale = 100,
                        coldists = distances, hsl = hsl_vals)

# Multiple images
snakes <- getimg(system.file("testdata/images/snakes", package = 'pavo'))
snakes_class <- classify(snakes, kcols = 3)
snakes_adj <- adjacent(snakes_class, xpts = 250, xscale = c(50, 55))
}

}
\references{
Endler, J. A. (2012). A framework for analysing colour pattern geometry:
adjacent colours. Biological Journal Of The Linnean Society, 86(4), 405-431.

Endler, J. A., Cole G., Kranz A.  (2018). Boundary Strength Analysis:
Combining color pattern geometry and coloured patch visual properties for use in predicting behaviour
and fitness. bioRxiv.
}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
