% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent.R
\name{adjacent}
\alias{adjacent}
\title{Run an adjacency analysis}
\usage{
adjacent(classimg, x_pts = NULL, x_scale = NULL, bkg_ID = NULL,
  bkg_include = TRUE, coldists = NULL, cores = getOption("mc.cores", 2L))
}
\arguments{
\item{classimg}{(required) an xyz image matrix, or list of matrices, in which
x and y correspond to pixel coordinates, and z is a numeric code specifying
a colour-class. Preferably the result of \code{\link{classify}}.}

\item{x_pts}{(required) an integer specifying the number of sample points, or grid-sampling
density, along the x axis. Y-axis sampling density is calculated automatically
from this, to maintain an even grid spacing.}

\item{x_scale}{(required) an integer specifying the true length of the x-axis,
in preferred units. Not required, and ignored, if image scales have been set via
\code{\link{calibrate}}.}

\item{bkg_ID}{an integer or vector specifying the colour-class ID number(s) of
pertaining to the background. Examine the attributes of, or call \code{summary} on,
the result of \code{\link{classify}} to visualise the RGB values corresponding to
colour-class ID numbers.}

\item{bkg_include}{logical; should the colour classes specified by \code{bkg_ID}
be excluded from the analyses? Defaults to \code{FALSE}. Note that if \code{TRUE}, ALL
members of the colour classes specified in \code{bkg_ID_i} will be excluded, which
may give undesired results if elements are shared between the background and focal
stimulus.}

\item{coldists}{A data.frame specifying the visually-modelled chromatic (dS)
and/or achromatic (dL) distances between colour-categories. The first two columns
should specify all possible combinations of colour category ID's, and be named 'c1'
and 'c2', with the remaining columns named dS (for chromatic distances) and/or dL
(for achromatic distances). See \code{\link{vismodel}} and \code{\link{colspace}}
for visual modelling with spectral data.}

\item{cores}{number of cores to be used in parallel processing. If \code{1}, parallel
computing will not be used. Defaults to \code{getOption("mc.cores", 2L)}.}
}
\value{
a data frame of summary variables:
\itemize{
  \item \code{'k'}: The number of user-specified colour and/or luminance classes.
  \item \code{'N'}: The grand total (sum of diagonal and off-diagonal) transitions.
  \item \code{'n_off'}: The total off-diagonal transitions.
  \item \code{'Obs_i_j'}: The observed number of off-diagonal (class-change)
    transitions, such that \code{sum(Obs_i_j) = n_off}.
  \item \code{'E_i_j'}: The expected number of off-diagonal (class-change)
    transitions, where \code{E_i_j = 2 x n_off x p_i x p_j}.
  \item \code{'d_t_o'}: The absolute deviation from expected off-diagonal (class-change)
    transitions, calculated on the observed transition matrix, where
    d_t_o = sum(abs(O_i_j - E_i_j)).
  \item \code{'p_i'}: The overall frequency of colour class \emph{i}.
  \item \code{'q_i_j'}: The frequency of transitions between \emph{all} colour classes
   \emph{i} and \emph{j}, such that \code{sum(q_i_j) = 1}.
  \item \code{'t_i_j'}: The frequency of off-diagonal (i.e. class-change
    transitions) transitions \emph{i} and \emph{j}, such that \code{sum(t_i_j) = 1}.
  \item \code{'m'}: The overall transition density (mean transitions),
    in units specified in the argument \code{x_scale}.
  \item \code{'m_r'}: The row transition density (mean row transitions),
    in user-specified units.
  \item \code{'m_c'}: The column transition density (mean column transitions),
    in user-specified units.
  \item \code{'A'}: The transition aspect ratio (< 1 = wide, > 1 = tall).
  \item \code{'B'}: The animal/background transition ratio.
  \item \code{'Sc'}: Simpson colour class diversity.
  \item \code{'St'}: Simpson transition diversity.
  \item \code{'Jc'}: Simpson colour class diversity relative to its achievable maximum.
  \item \code{'Jt'}: Simpson transition diversity relative to its achievable maximum.
  \item \code{'m_dS'}: weighted mean of the chromatic edge magnitude
  \item \code{'s_dS'}: weighted standard deviation of the chromatic edge magnitude
  \item \code{'cv_dS'}: weighted coefficient of variation of the chromatic edge magnitude
  \item \code{'m_dL'}: weighted mean of the achromatic edge magnitude
  \item \code{'s_dL'}: weighted standard deviation of the achromatic edge magnitude
  \item \code{'cv_dL'}: weighted coefficient of variation of the achromatic edge magnitude
  }
}
\description{
Calculate summary variables from the adjacency analysis (Endler 2012 & Endler et al. 2018).
}
\examples{
\dontrun{
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))
papilio_class <- classify(papilio, n_cols = 4)
papilio_adj <- adjacent(papilio_class, x_pts = 150, x_scale = 100)

# Multiple images
snakes <- getimg(system.file("testdata/images/snakes", package = 'pavo'))
snakes_class <- classify(snakes, n_cols = 3)
snakes_adj <- adjacent(snakes_class, x_pts = 250, x_scale = 50)
}

}
\references{
Endler, J. A. (2012). A framework for analysing colour pattern geometry:
adjacent colours. Biological Journal Of The Linnean Society, 86(4), 405-431.

Endler, J. A., Cole G., Kranz A.  (2018). Combining color pattern
geometry and coloured patch visual properties for use in predicting behaviour
and fitness. Methods in Ecology and Evolution, Early View.
}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
