% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent.R
\name{adjacent_main}
\alias{adjacent_main}
\title{Main function for adjacency analysis}
\usage{
adjacent_main(classimg_i, xpts_i = NULL, xscale_i = NULL, bkgID_i = NULL,
  exclude2_i = TRUE, coldists_i = NULL)
}
\arguments{
\item{classimg_i}{(required) an xyz image matrix, or list of matrices, in which
x and y correspond to pixel coordinates, and z is a numeric code specifying
a colour-class. Preferably the result of \code{\link{classify}}.}

\item{xpts_i}{(required) an integer specifying the number of sample points, or grid-sampling
density, along the x axis. Y-axis sampling density is calculated automatically
from this, to maintain an even grid spacing.}

\item{xscale_i}{(required) an integer specifying the true length of the x-axis,
in preferred units. Not required, and ignored, if image scales have been set via
\code{\link{procimg}}.}

\item{bkgID_i}{an integer or vector specifying the colour-class ID number(s) of
pertaining to the background. Examine the attributes of, or call \code{summary} on,
the result of \code{\link{classify}} to visualise the RGB values corresponding to
colour-class ID numbers.}

\item{exclude2_i}{The portion of the image to be excluded from the analysis, if any.
If excluding the focal object, its outline must first have been idenfitied using 
\code{\link{procimg}}. If excluding the image background it must either have been
identified using \code{\link{procspec}}, or if it is relatively homogeneous, 
the colour-class ID's corresponding to the background can be specified using 
bkgID.}

\item{coldists_i}{An data.frame or matrix specifying the visually-modelled chromatic (dS)
and/or achromatic (dL) distances between colour-categories. The first two columns
should specify all possible combinations of colour category ID's, with the remaining
columns named dS (for chromatic distances) and/or dL (for achromatic distances).}
}
\description{
Main function for adjacency analysis
}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
\keyword{internal}
