% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent.R
\name{adjacent_main}
\alias{adjacent_main}
\title{Main function for adjacency analysis}
\usage{
adjacent_main(classimg_i, x_pts_i = NULL, x_scale_i = NULL,
  bkg_ID_i = NULL, bkg_include_i = TRUE, coldists_i = NULL)
}
\arguments{
\item{classimg_i}{(required) an xyz image matrix, or list of matrices, in which
x and y correspond to pixel coordinates, and z is a numeric code specifying
a colour-class. Preferably the result of \code{\link{classify}}.}

\item{x_pts_i}{(required) an integer specifying the number of sample points, or grid-sampling
density, along the x axis. Y-axis sampling density is calculated automatically
from this, to maintain an even grid spacing.}

\item{x_scale_i}{(required) an integer specifying the true length of the x-axis,
in preferred units. Not required, and ignored, if image scales have been set via
\code{\link{calibrate}}.}

\item{bkg_ID_i}{an integer specifying the colour-class ID number of the homogeneous background.
Examine the attributes of, or call \code{summary} on, the result of \code{\link{classify}}
to visualise the RGB values corresponding to colour-class ID numbers.}

\item{bkg_include_i}{logical; should the background be excluded from the analyses?
Defaults to \code{FALSE}.}

\item{coldists_i}{An data.frame or matrix specifying the visually-modelled chromatic (dS)
and/or achromatic (dL) distances between colour-categories. The first two columns
should specify all possible combinations of colour category ID's, with the remaining
columns named dS (for chromatic distances) and/or dL (for achromatic distances).}
}
\value{
a data frame of summary variables. See \code{\link{adjacent}} for details.
}
\description{
Main function for adjacency analysis
}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
\keyword{internal}
