% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate images}
\usage{
calibrate(image, scale_length = NULL)
}
\arguments{
\item{image}{(required) image data. Either a single image array, or a number of images
stored in a list. Preferably the result of \code{\link{getimg}}.}

\item{scale_length}{an integer specifying the length of the scale
in the image(s), if desired.}
}
\value{
an image array, or list containing images, for use in further
\code{pavo} functions, with scales stored as an attribute.
}
\description{
Manually specify the scale of images.
}
\examples{
\dontrun{
# Single image
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))
papilio <- calibrate(papilio, scale = 10)

# Multiple images. Assign a scale an specify the nuber of colours present.
snakes <- getimg(system.file("testdata/images/snakes", package = 'pavo'))
snakes <- calibrate(snakes, scale = 100, assign_n = TRUE)
}

}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
