% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical}
\alias{categorical}
\title{Categorical fly-visual model}
\usage{
categorical(vismodeldata)
}
\arguments{
\item{vismodeldata}{(required) quantum catch color data. Can be either the result
from \code{\link{vismodel}} or independently calculated data (in the form of a data frame
with four columns representing a tetrachromatic dipteran viewer).}
}
\value{
Object of class \code{colspace} consisting of the following columns:

\code{R7p, R7y, R8p, R8y}: the quantum catch data used to
 calculate the remaining variables.

\code{x, y}: cartesian coordinates in the categorical colour space.

\code{category}: fly-colour category. One of \code{p-y-}, \code{p-y+}, \code{p+y-}, \code{p+y+}.
}
\description{
Applies the fly categorical colour vision model of Troje (1993)
}
\examples{
\dontrun{
data(flowers)
vis.flowers <- vismodel(flowers, qcatch = 'Qi', visual = 'musca', achro = 'none', relative = TRUE)
cat.flowers <- colspace(vis.flowers, space = 'categorical')
}

}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\references{
Troje N. (1993). Spectral categories in the learning behaviour
 of blowflies. Zeitschrift fur Naturforschung C, 48, 96-96.
}
\keyword{internal}

