% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cieplot.R
\name{cieplot}
\alias{cieplot}
\title{CIE plot}
\usage{
cieplot(ciedata, mono = TRUE, out.lwd = NULL, out.lcol = "black",
  out.lty = 1, ...)
}
\arguments{
\item{ciedata}{(required)}

\item{mono}{should the monochromatic loci (the 'horseshoe') be
plotted? Defaults to \code{TRUE}}

\item{out.lwd}{line width for monochromatic loci outline (defaults to 1)}

\item{out.lcol}{line colour for monochromatic loci outline (defaults to black)}

\item{out.lty}{line type for monochromatic loci outline (defaults to 1)}

\item{...}{Additional graphical options. See \code{\link{par}}.}
}
\description{
Plot a CIE (XYZ or LAB) chromaticity diagram. Accessed via the function \code{\link{plot.colspace}}.
}
\examples{
\dontrun{
data(flowers)
vis.flowers <- vismodel(flowers, visual = 'cie2', illum = 'D65')
flowers.cie <- colspace(vis.flowers, space = 'ciexyz')
plot(flowers.cie)
}

}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\references{
Smith T, Guild J. (1932) The CIE colorimetric standards and their use.
   Transactions of the Optical Society, 33(3), 73-134.

Westland S, Ripamonti C, Cheung V. (2012). Computational colour science 
   using MATLAB. John Wiley & Sons.
}
\keyword{internal}

