% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Identify colour classes in an image for adjacency analyses}
\usage{
classify(imgdat, n_cols, ref_ID = NULL, manual = FALSE,
  cores = getOption("mc.cores", 2L))
}
\arguments{
\item{imgdat}{(required) image data. Either a single image, or a series of images
stored in a list. preferably the result of \code{\link{getimg}}.}

\item{n_cols}{(required) either an integer, or vector the same length as imgdat (if
passing a list of images), specifying the number of discrete colour classes present
in an image, for k-means clustering. Ignored if n_cols has already been set via
\code{\link{calibrate}}.}

\item{ref_ID}{The numeric identifier of the 'reference' image, for use when passing
a list of images. Other images will be k-means classified using centres identified
in the reference image, thus helping to ensure that homologous pattern elements
will be reliably classified between images. Defaults to the first image in the list.
Ignored if n_cols is a vector.}

\item{manual}{manually specify the colour-category 'centers', for k-means clustering.
When \code{TRUE}, the user is asked to click a number of points (equal to \code{n_cols})
on the reference image that represent the distinct colours of interest.}

\item{cores}{number of cores to be used in parallel processing. If \code{1}, parallel
computing will not be used. Defaults to \code{getOption("mc.cores", 2L)}.}
}
\value{
A matrix, or list of matrices, of class \code{rimg} containing the colour
class classifications at each pixel location. The RGB values corresponding to
k-means centres (i.e. colour classes) are stored as object attributes.
}
\description{
Use k-means clustering to classify image pixels into discrete colour classes.
}
\note{
Since the \code{kmeans} process draws on random numbers to find initial
cluster centres when \code{manual = FALSE}, use \code{set.seed} if reproducible
cluster ID's are desired between runs.
}
\examples{
\dontrun{
# Single image
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))
papilio_class <- classify(papilio, n_cols = 4)

# Multiple images
snakes <- getimg(system.file("testdata/images/snakes", package = 'pavo'))
snakes_class <- classify(snakes, n_cols = 3)
}

}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
