% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldist.R
\name{coldist}
\alias{coldist}
\title{Color distances}
\usage{
coldist(qdata, noise = c("neural", "quantum"), subset = NULL,
  achro = TRUE, vis = NULL, qcatch = NULL, n1 = 1, n2 = 2, n3 = 2,
  n4 = 4, weber = 0.1, weber.ref = "n4", weber.achro = 0.1)
}
\arguments{
\item{qdata}{(required) quantum catch color data. Can be the result
from \code{\link{vismodel}}, or \code{\link{colspace}}. Data may also be independently calculated quantum catches, 
in the form of a data frame with columns representing photoreceptors.}

\item{noise}{how the noise will be calculated. (Ignored for \code{colspace} objects
 if model is not a receptor noise model (i.e. hexagon, colour-opponent-coding, 
categorical, and cielab models)):
\itemize{
    \item \code{neural}: noise is proportional to the Weber fraction and is independent of the
    intensity of the signal received.
    \item \code{quantum}: noise is the sum of the neural noise and receptor noise, and is thus
    proportional to the Weber fraction and inversely proportional to the intensity of the signal
    received (the quantum catches). Note that the \code{quantum} option will only work with 
    objects of class \code{vismodel}.
}}

\item{subset}{If only some of the comparisons should be returned, a character vector of 
length 1 or 2 can be provided, indicating which samples are desired. The subset vector 
must match the labels of the imput samples, but partial matching (and regular expressions) 
are supported.}

\item{achro}{Logical. If \code{TRUE}, last column of the data frame is used to calculate 
the achromatic contrast, with noise based on the Weber fraction calculated using \code{n4}. 
If the data are from the hexagon model (i.e. \code{colspace(space = 'hexagon')}), it 
instead returns long (or 'green') receptor contrast.}

\item{vis}{if the object is of class \code{vismodel} or \code{colspace}, 
 this argument is ignored. If the object is a data frame of quantal catches 
 from another source, this argument is used to specify the visual system phenotype 
 to use in the model:
\itemize{
 \item \code{tcs}: Tetrachromatic color vision (default)
 \item \code{tri}: Trichromatic color vision
 \item \code{di}: Dichromatic color vision
}}

\item{qcatch}{if the object is of class \code{vismodel} or \code{colspace}, 
 this argument is ignored. If the object is a data frame of quantal catches 
 from another source, this argument is used to specify what type of quantum catch is being 
 used, so that the noise can be calculated accordingly: 
\itemize{
 \item \code{Qi}: Quantum catch for each photoreceptor
 \item \code{fi}: Quantum catch according to Fechner law (the signal of the receptor
   channel is proportional to the logarithm of the quantum catch)
 }}

\item{n1, n2, n3, n4}{tetrachromatic photoreceptor densities for u, s, m & l (default to 
the Pekin robin \emph{Leiothrix lutea} densities: 1:2:2:4). If \code{vis} does not equal 
\code{'tcs'}, only \code{n1} and \code{n2} (\code{vis = 'di'}) or 
\code{n1}, \code{n2} and \code{n3} (\code{vis = 'tri'}) are used for chromatic contrast. 
Ignored when data are of class \code{\link{colspace}}.}

\item{weber}{The Weber fraction to be used. The noise-to-signal ratio \code{v} is unknown, 
 and therefore must be calculated based on the epirically estimated Weber fraction of one of
 the cone classes. \code{v} is then applied to estimate the Weber fraction of the 
 other cones. by default, the value of 0.1 is used (the empirically estimated value for the
 LWS cone from \emph{Leiothrix lutea}). Ignored for \code{colspace} objects
 if model is not a receptor noise model (i.e. hexagon, colour-opponent-coding, 
categorical, and cielab models).}

\item{weber.ref}{the cone class used to obtain the empirical estimate of the 
 Weber fraction used for the \code{weber} argument. By default, \code{n4} is used, 
 representing the LWS cone for \emph{Leiothrix lutea}. Ignored for \code{colspace} objects
 if model is not a receptor noise model (i.e. hexagon, colour-opponent-coding, 
categorical, and cielab models).}

\item{weber.achro}{the Weber fraction to be used to calculate achromatic contrast, when 
 \code{achro = TRUE}. Defaults to 0.1. Ignored for \code{colspace} objects
 if model is not a receptor noise model (i.e. hexagon, colour-opponent-coding, 
categorical, and cielab models).}
}
\value{
A data frame containing up to 4 columns. The first two (\code{patch1, patch2}) refer
to the two colors being contrasted; \code{dS} is the chromatic contrast (delta S)
and \code{dL} is the achromatic contrast (delta L). Units are JND's in the receptor-noise
model, euclidean distances in the hexagon and cielab colorspaces, and manhattan distances
in the color-opponent-coding space.
}
\description{
Calculates color distances. When data are the result of \code{\link{vismodel}}, 
it applies the receptor-noise model of Vorobyev et al. (1998) to calculate color distances
with noise based on relative photoreceptor densities. It also accepts \code{\link{colspace}} data 
from the hexagon, colour-opponent-coding, categorical, and cielab models, in which case euclidean
distances (hexagon, cielab, categorical) or manhattan distances (coc) are returned.
}
\examples{
\dontrun{
# Dichromat
data(flowers)
vis.flowers <- vismodel(flowers, visual = 'canis', relative= FALSE)
didist.flowers <- coldist(vis.flowers)

# Trichromat 
vis.flowers <- vismodel(flowers, visual = 'apis', relative = FALSE)
tridist.flowers <- coldist(vis.flowers)

# Trichromat, color-hexagon model
vis.flowers <- vismodel(flowers, visual = 'apis', qcatch = 'Ei', relative = FALSE, vonkries = TRUE, achro = 'l', bkg = 'green')
hex.flowers <- colspace(vis.flowers, space = 'hexagon')
hexdist.flowers <- coldist(hex.flowers)

# Tetrachromat
data(sicalis)
vis.sicalis <- vismodel(sicalis, visual = 'avg.uv', relative = FALSE)
tetradist.sicalis.n <- coldist(vis.sicalis)

# this will also work, but give you several warnings you shouldn't ignore!!
col.sicalis <- colspace(vis.sicalis)
tetradist.sicalis.n <- coldist(col.sicalis)

tetradist.sicalis.q <- coldist(vis.sicalis, noise='quantum')
}

}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}
}
\references{
Vorobyev, M., Osorio, D., Bennett, A., Marshall, N., & Cuthill, I. 
 (1998). Tetrachromacy, oil droplets and bird plumage colours. Journal Of Comparative 
 Physiology A-Neuroethology Sensory Neural And Behavioral Physiology, 183(5), 621-633.

Hart, N. S. (2001). The visual ecology of avian photoreceptors. Progress 
 In Retinal And Eye Research, 20(5), 675-703.

Endler, J. A., & Mielke, P. (2005). Comparing entire colour patterns 
 as birds see them. Biological Journal Of The Linnean Society, 86(4), 405-431.
}

