% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-flowers.R
\docType{data}
\name{flowers}
\alias{flowers}
\title{Reflectane spectra from a suite of native Australian flowers, collected
around Cairns, Queensland.}
\description{
dataset containing reflectance measurements from 36 native 
Australian angiosperm species, indicated by column names.
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\keyword{data}

