% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jnd2xyz.R
\name{jnd2xyz}
\alias{jnd2xyz}
\title{Convert JND distances into perceptually-corrected Cartesian coordinates}
\usage{
jnd2xyz(coldistres)
}
\arguments{
\item{coldistres}{(required) the output from a \code{coldist} call.}
}
\description{
Converts a \code{coldist} output into Cartesian coordinates that are
perceptually-corrected (i.e. Euclidean distances = JND distances)
}
\examples{
\dontrun{
data(flowers)
vis.flowers <- vismodel(flowers)
cd.flowers <- coldist(vis.flowers)
jnd2xyz(cd.flowers)
}

}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}
}
