% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxplot.R
\name{maxplot}
\alias{maxplot}
\title{Plot a Maxwell Triangle}
\usage{
maxplot(maxdata, labels = TRUE, achro = TRUE, achrocol = "grey",
  achrosize = 0.8, cex.labels = 1, out.lwd = 1, out.lcol = "black",
  out.lty = 1, ...)
}
\arguments{
\item{maxdata}{(required) a data frame, possibly a result from the \code{maxwell}
function, containing values for the 'x' and 'y' coordinates as columns (labeled as such)}

\item{labels}{plot verticy labels? Defaults to \code{TRUE}}

\item{achro}{should a point be plotted at the origin (defaults to \code{TRUE})?}

\item{achrocol}{color of the point at the origin \code{achro = TRUE} (defaults to grey)}

\item{achrosize}{size of the point at the origin when \code{achro = TRUE} (defaults to 0.8)}

\item{cex.labels}{character expansion factor for category labels when \code{labels = TRUE})}

\item{out.lwd}{line width for triangle outline (defaults to 1)}

\item{out.lcol}{line colour for triangle outline (defaults to black)}

\item{out.lty}{line type for triangle outline (defaults to 1)}

\item{...}{additional graphical options. See code{\link{par}}.}

\item{labsize}{size of the text labels (defaults to 1)}
}
\description{
\code{maxplot} produces a Maxwell triangle plot
}
\examples{
\dontrun{
data(flowers)
vis.flowers <- vismodel(flowers, visual = 'apis')
maxwell.flowers <- maxwell(vis.flowers)
maxplot(maxwell.flowers)
}
}
\author{
Thomas White \email{thomas.white026@gmail.com}
}
\references{
Kelber A, Vorobyev M, Osorio D. (2003). Animal colour vision
   - behavioural tests and physiological concepts. Biological Reviews, 78,
   81 - 118.

Neumeyer C (1980) Simultaneous color contrast in the honeybee.
 Journal of comparative physiology, 139(3), 165-176.
}

