% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pavo-package.R
\docType{package}
\name{pavo-package}
\alias{pavo}
\alias{pavo-package}
\title{pavo: Perceptual Analysis, Visualization and Organization of Spectral Color
Data in R}
\description{
A cohesive framework for parsing, analyzing and organizing color from spectral data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://rafaelmaia.net/pavo/}
}

}
\author{
\strong{Maintainer}: Rafael Maia \email{rm72@zips.uakron.edu}

Authors:
\itemize{
  \item Thomas White \email{thomas.white026@gmail.com}
  \item Hugo Gruson \email{hugo.gruson+R@normalesup.org}
  \item Chad Eliason \email{cme16@zips.uakron.edu}
  \item Pierre-Paul Bitton \email{bittonp@uwindsor.ca}
}

}
\keyword{internal}
