% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.colorspace.r
\name{plot.colorspace}
\alias{plot.colorspace}
\title{Plot spectra in a colourspace}
\usage{
\method{plot}{colorspace}(clrspdata, ...)
}
\arguments{
\item{clrspdata}{(required) an object of class \code{colorspace}. May be a result from
\code{dispace}, \code{maxwell}, \code{coc}, \code{hexagon}, \code{categorical}, 
or \code{tcs}. Note that the below options will vary by plot.}

\item{...}{additional graphical options. See code{\link{par}}}

\item{labels}{plot verticy labels (or category labels, when clrspdata is the result of
\code{categorical})? Defaults to \code{TRUE}}

\item{cex.labels}{character expansion factor for labels when \code{labels = TRUE})}

\item{achro}{should a point be plotted at the origin (defaults to \code{TRUE})? Has
no effect when clrspdata is the result of \code{categorical} or \code{coc}}

\item{achrosize}{size of the point at the origin when \code{achro = TRUE} (defaults to 0.8)}

\item{achrocol}{color of the point at the origin \code{achro = TRUE} (defaults to grey)}

\item{out.lwd}{line width for plot outline (defaults to 1). Has no effect when
clrspdata is the result of \code{coc}}

\item{out.lcol}{line colour for plot outline (defaults to black). Has no effect when
clrspdata is the result of \code{coc}}

\item{out.lty}{line type for plot outline (defaults to 1). Has no effect when
clrspdata is the result of \code{coc}}

\item{tick.loc}{a numeric vector specifying the location of tick marks on x & y axes,
only when clrspdata is the result of \code{dispace}}

\item{sectors}{plot the bee-hue sector dividers (only when clrspdata is the result 
of \code{hexagon})? Options are:
   \itemize{ 
       \item \code{'none'}: No sectors (default)
       \item \code{'fine'}: 36 10-degree sectors
       \item \code{'coarse'}: six bee-hue sectors (blue, blue-green, green, uv-green, uv, uv-blue).
       }}

\item{col.sec}{line colour of bee-hue sector dividers. Defaults to \code{'grey'}}
}
\description{
\code{plot.colorspace} Plots reflectance spectra in the appropriate colorspace
}
\examples{
\dontrun{
data(flowers)

# Dichromat space
vis.flowers <- vismodel(flowers, visual = 'canis')
di.flowers <- dispace(vis.flowers)
plot(di.flowers)

# Colour hexagon 
vis.flowers <- vismodel(flowers, visual = 'apis', qcatch = 'Ei', relative = FALSE, vonkries = TRUE, achro = 'l', bkg = 'green')
hex.flowers <- hexagon(vis.flowers)
plot(hex.flowers)

# Maxwell triangle
vis.flowers <- vismodel(flowers, visual = 'apis')
maxwell.flowers <- maxwell(vis.flowers)
plot(maxwell.flowers)
}
 
}
\seealso{
\code{\link{plot}}
}

