% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.vismodel}
\alias{plot.vismodel}
\title{New colorspace plotter}
\usage{
\method{plot}{vismodel}(vismodeldat, achro.pt = FALSE, achro.line = FALSE,
  hull = FALSE, zoom = 1, view = c("tetrahedron", "crop"),
  type = c("hexagon", "cie", "macleod"), ...)
}
\arguments{
\item{vismodeldata}{(required) quantum catch color data. Can be either the result
of \code{\link{vismodel}} or independently calculated data in the form of a data frame
with columns representing the quantum catches of different cones/receptors.}

\item{achro.pt}{whether to plot the achromatic center on the chromaticity diagram}

\item{achro.line}{whether to add a line in 3D tetrahedral plots representing the
achromatic center}

\item{hull}{whether to add convex hulls around data points}

\item{zoom}{zoom factor to multiple data points by (?)}

\item{view}{specifies whether the plot should contain the entire chromaticity diagram
(default) or only the given data points}

\item{type}{for visual models with <4 cones, specifies the type of chromaticity
diagram to produce (defaults to Chittka's hexagon diagram)}
}
\value{
a chromaticity diagram
}
\description{
DESCRIPTION HERE...
}
\examples{
\dontrun{
data(sicalis)
vis.sicalis <- vismodel(sicalis, visual='avg.uv')
tcs.sicalis <- tcs(vis.sicalis)}
plot3d(tcs(vismodel(teal)), theta=60, phi=25)
}
\author{
Chad Eliason \email{cme16@zips.uakron.edu}

Rafael Maia \email{rm72@zips.uakron.edu}
}
\references{
Stoddard, M. C., & Prum, R. O. (2008). Evolution of avian plumage color in a tetrahedral color space: A phylogenetic analysis of new world buntings. The American Naturalist, 171(6), 755-776.

Endler, J. A., & Mielke, P. (2005). Comparing entire colour patterns as birds see them. Biological Journal Of The Linnean Society, 86(4), 405-431.
}

