% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procimg.R
\name{procimg}
\alias{procimg}
\title{Process images}
\usage{
procimg(image, scaledist = NULL, outline = FALSE, smooth = FALSE,
  iterations = 1L, plotnew = FALSE, ...)
}
\arguments{
\item{image}{(required) image data. Either a single image array, or a number of images
stored in a list. Preferably the result of \code{\link{getimg}}.}

\item{scaledist}{an integer, or numeric vector equal in length to the number of images,
specifying the length of the scale in the image(s).}

\item{outline}{allows the user to interactively specify the focal object in
an image by clicking around its outline. The xy-coordinates of the resulting
polygon are saved as an attribute, for use in genrating a masking layer &
separating animals/plants from backgrounds in further analyses. This is particularly
useful when backgrounds are complex, (e.g. in natural settings) in which case
backgrounds and objects cannot be readily separated by simple k-means clustering.}

\item{smooth}{should the polygon specified when \code{outline = TRUE} be smoothed
using Chaikin's corner-cuting algorithm? Defaults to \code{FALSE}.}

\item{iterations}{the number of smoothing iterations, when \code{smooth = TRUE}.}

\item{plotnew}{should plots be opened in a new window? Defaults to \code{FALSE}.}

\item{...}{additional graphical parameters. Also see \code{\link{par}}.}
}
\value{
an image array, or list containing images, for use in further
\code{pavo} functions, with scales stored as an attribute.
}
\description{
Manually specify the scale of images.
}
\examples{
\dontrun{
# Single image
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))
papilio <- procimg(papilio, scaledist = 10)

# Assign individual scales to each image.
snakes <- getimg(system.file("testdata/images/snakes", package = 'pavo'))
snakes <- procimg(snakes, scaledist = c(10, 14))
}

}
\references{
Chaikin, G. 1974. An algorithm for high speed curve generation.
Computer Graphics and Image Processing 3, 346-349.
}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
