% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procspec.R
\name{procspec}
\alias{procspec}
\title{Process spectra}
\usage{
procspec(
  rspecdata,
  opt = c("none", "smooth", "maximum", "minimum", "bin", "sum", "center", "clip"),
  fixneg = c("none", "addmin", "zero"),
  span = 0.25,
  bins = 20,
  clip_range = NULL
)
}
\arguments{
\item{rspecdata}{(required) a data frame, possibly of class \code{rspec}, which
contains a column containing a wavelength range, named 'wl', and spectra
data in remaining columns.}

\item{opt}{what type of processing options to apply. User can select multiple
options by providing a vector. Possibilities are:
\itemize{
\item \code{"none"} does not perform any processing (default).
\item \code{"smooth"} applies LOESS smoothing to each spectrum using \code{\link[=loess.smooth]{loess.smooth()}}.
Optimal smoothing parameter can be assessed by using \code{\link[=plotsmooth]{plotsmooth()}}.
\item \code{"minimum"} subtracts the minimum from each individual spectra.
\item \code{"maximum"} divides each spectrum by its maximum value.
\item \code{"sum"} divides each spectrum by summed values.
\item \code{"bin"} bins each spectrum into the specified number of bins. \code{bins}
argument must be set.
\item \code{"center"} centers individual spectra by subtracting mean reflectance from
all values.
\item \code{"clip"} removes a specified range of wavelengths and replaces them by
linear interpolation (clipping occurs before smoothing). \code{clip_range} must be provided.
}}

\item{fixneg}{how to handle negative values. Possibilities are:
\itemize{
\item \code{"none"} does not perform negative value correction (default).
\item \code{"zero"} sets all negative values to zero.
\item \code{"addmin"} adds the absolute value of the maximally negative values of each
spectra to the reflectance at all other wavelengths (setting the minimum
value to zero, but scaling other values accordingly).
}}

\item{span}{sets the smoothing parameter used by \code{\link[=loess.smooth]{loess.smooth()}}.}

\item{bins}{sets the number of equally sized wavelength bins for \code{opt = "bin"}.}

\item{clip_range}{either a numeric vector indicating the two bounds of the range
of wavelengths to clip for \code{opt = "clip"}, or a list of such numeric vectors if multiple ranges
are to be clipped.}
}
\value{
A data frame of class \code{rspec} with the processed data.
}
\description{
Applies normalization and/or smoothing to spectra for further analysis or plotting.
}
\examples{
data(teal)
plot(teal, select = 10)

# Smooth data to remove noise
teal.sm <- procspec(teal, opt = "smooth", span = 0.25)
plot(teal.sm, select = 10)

# Normalize to max of unity
teal.max <- procspec(teal, opt = c("max"))
plot(teal.max, select = 10)

# Smoothing directly severe artifacts can artificially modify the shape of
# the entire spectrum. In this case, it is better to clip the artifact and
# then smooth the data.
teal_clip <- procspec(teal, opt = c("clip", "smooth"), clip_range = c(600, 650), span = 0.25)
plot(teal_clip, select = 10)

}
\references{
Cuthill, I., Bennett, A. T. D., Partridge, J. & Maier, E. 1999.
Plumage reflectance and the objective assessment of avian sexual dichromatism.
The American Naturalist, 153, 183-200.

Montgomerie R. 2006. Analyzing colors. In Hill, G.E, and McGraw,
K.J., eds. Bird Coloration. Volume 1 Mechanisms and measurements. Harvard
University Press, Cambridge, Massachusetts.

White, T. E., Dalrymple, R. L., Noble D. W. A., O'Hanlon, J. C.,
Zurek, D. B., Umbers, K. D. L. 2015. Reproducible research in the study
of biological coloration. Animal Behaviour, 106, 51-57.
}
\seealso{
\code{\link[=loess.smooth]{loess.smooth()}}, \code{\link[=plotsmooth]{plotsmooth()}}
}
\author{
Chad Eliason \email{cme16@zips.uakron.edu}
}
