% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensdata.R
\name{sensdata}
\alias{sensdata}
\title{Plot or Retrieve In-built Data}
\usage{
sensdata(visual = c("none", "avg.uv", "avg.v", "bluetit", "star", "pfowl",
  "apis", "canis", "cie2", "cie10", "musca", "segment"),
  achromatic = c("none", "bt.dc", "ch.dc", "st.dc", "md.r1"),
  illum = c("none", "bluesky", "D65", "forestshade"), trans = c("none",
  "bluetit", "blackbird"), bkg = c("none", "green"), plot = TRUE, ...)
}
\arguments{
\item{visual}{visual systems. Options are:
\itemize{
\item \code{none}: no sensitivity data
\item \code{apis}: Honeybee \emph{Apis mellifera} visual system.
\item \code{avg.uv}: average avian UV system.
\item \code{avg.v}: average avian V system.
\item \code{bluetit}: Blue tit \emph{Cyanistes caeruleus} visual system.
\item \code{canis}: Canid \emph{Canis familiaris} visual system.
\item \code{cie2}: 2-degree colour matching functions for CIE models of human 
 colour vision. Functions are linear transformations of the 2-degree cone fundamentals 
 of Stockman & Sharpe (2000), as ratified by the CIE (2006).
\item \code{cie10}: 10-degree colour matching functions for CIE models of human 
 colour vision. Functions are linear transformations of the 10-degree cone fundamentals 
 of Stockman & Sharpe (2000), as ratified by the CIE (2006).
\item \code{musca}: Housefly \emph{Musca domestica} visual system.
\item \code{pfowl}: Peafowl \emph{Pavo cristatus} visual system.
\item \code{segment}: Generic tetrachromat 'viewer' for use in the segment analysis of Endler (1990).
\item \code{star}: Starling \emph{Sturnus vulgaris} visual system.
}}

\item{achromatic}{the sensitivity data used to calculate luminance (achromatic)
 receptor stimulation. Options are: 
\itemize{
 \item \code{none}: no achromatic sensitivity data
\item \code{bt.dc}: Blue tit \emph{Cyanistes caeruleus} double cone
 \item \code{ch.dc}: Chicken \emph{Gallus gallus} double cone
 \item \code{st.dc}: Starling \emph{Sturnus vulgaris} double cone
 \item \code{md.r1}: Housefly \emph{Musca domestica} R1-6 photoreceptor
}}

\item{illum}{illuminants. Options are:
\itemize{ 
\item \code{none}: no illuminant data
\item \code{'bluesky'} open blue sky
\item \code{'D65'}: standard daylight
\item \code{'forestshade'} forest shade
}}

\item{trans}{Ocular transmission data. Options are:
\itemize{ 
\item \code{none}: no transmission data
\item \code{'bluetit'}: blue tit \emph{Cyanistes caeruleus} 
ocular transmission (from Hart et al. 2000)
\item \code{'blackbird'}: blackbird \emph{Turdus merula} 
ocular transmission (from Hart et al. 2000)
}}

\item{bkg}{background spectra. Options are:
\itemize{ 
\item \code{none}: no background spectral data
\item \code{'green'}: green foliage
}}

\item{plot}{should the data be plotted, or returned instead (defaults to \code{TRUE})?}

\item{...}{additional graphical options passed to \code{\link{plot.rspec}} when \code{plot = TRUE}.}
}
\description{
Plots or retrieves pavo's in-built spectral data.
}
\examples{
\dontrun{
# Blue tit's receptor
sensdata(visual = 'bluetit', achromatic = 'bt.dc', ylab = 'Absorbance')

# Average UV vs V avian receptors
sensdata(visual = c('avg.v', 'avg.uv'), ylab = 'Absorbance')

# CIE colour matching functions
sensdata(visual = c('cie2', 'cie10'), ylab = 'Absorbance')
}

}
\author{
Thomas White \email{thomas.white026@gmail.com}

Rafael Maia \email{rm72@zips.uakron.edu}
}
