% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{sicalis}
\alias{sicalis}
\title{Spectral curves from three body regions of Stripe-Tailed Yellow Finch
(\emph{Sicalis citrina}) males}
\description{
dataset containing reflectance measurements from 3 body parts
("C": crown, "B": breast, "T": throat) from seven male stripe-tailed yellow finches
}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}
}
\keyword{data}

