% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spec.R
\name{simulate_spec}
\alias{simulate_spec}
\title{Simulate a reflectance spectrum}
\usage{
simulate_spec(
  wl_inflect = NULL,
  wl_peak = NULL,
  width_sig = 20,
  width_gauss = 70,
  xlim = c(300, 700),
  ylim = c(0, 100)
)
}
\arguments{
\item{wl_inflect}{A numeric vector specifying the wavelength location (in nm)
for one or more inflection point(s) for a 'sigmoid' shaped curve, if desired.}

\item{wl_peak}{A numeric vector specifying the wavelength location (in nm)
for one or more inflection point(s) for a 'Gaussian' (or 'bell') shaped curve,
if desired.}

\item{width_sig}{A numeric value or vector (if multiple wl_inflect values are specified)
specifying the steepness of the change, for any sigmoidal curve(s). Required when
\code{wl_peak} is specified. Defaults to 20 nm.}

\item{width_gauss}{A numeric value or vector specifying the the full-width at half-maximum
of any Gaussian curve(s). Required when \code{wl_peak} is specified. Defaults to 70 nm.}

\item{xlim}{A vector specifying the wavelength range of the simulated spectra. Defaults to
300-700nm (\code{c(300, 700)}).}

\item{ylim}{A vector specifying the minimum and maximum reflectance values of the resulting
curve. Defaults to 0 - 100 \% (\code{c(0, 100)}).}
}
\value{
An object of class \code{rspec}.
}
\description{
Simulate a natural reflectance spectrum with sigmoidal (s-shaped) and/or Gaussian
(bell-shaped) curves. Multiple Gaussian and sigmoidal curves can be combined in
a single spectrum, to simulate more complex reflectance functions.
}
\examples{
# Single ideal 'grey' reflectance spectrum, with 50\% reflectance across 300 - 700 nm.
spec0 <- simulate_spec(ylim = c(0, 50))

# Single sigmoidal spectrum, with a low-to-high inflection at 550 nm.
spec1 <- simulate_spec(wl_inflect = 550)

# Single Gaussian spectrum, with a peak at 400 nm
spec2 <- simulate_spec(wl_peak = 400)

# Combination of both Gaussian (with peak at 340 nm) and sigmoidal (with inflection
# at 560 nm)
spec3 <- simulate_spec(wl_inflect = 560, wl_peak)

# Double-Gaussian peaks of differing widths
spec4 <- simulate_spec(wl_peak = c(340, 560), width_gauss = c(12, 40))

# Complex spectrim with single sigmoidal peak and multi-Gaussian peaks
spec5 <- simulate_spec(wl_inflect = 575, wl_peak = c(340, 430), width_gauss = c(20, 60))

# Simulate a set of Gaussian reflectance curves with peaks varying between 400-600 in 
10 nm increments, then combine into a single rspec object, and plot the result
peaks <- seq(400, 600, 10)  # Peak locations
spectra <- lapply(seq_along(peaks), function(x) simulate_spec(wl_peak = peaks[x]))  # Simulate
spectra <- Reduce(function(...) merge(..., all = TRUE), spectra)  # Combine
plot(spectra)  # Plot

Simulate a set of Gaussian reflectance curves with a single peak at 500 nm, but
with maximum reflectance varying from 30 to 90\% in 10\% increments, then combine 
into a single rspec object, and plot the result
ymax <- lapply(seq(30, 90, 10), function(x) c(0, x))  # Varying reflectance maxima
spectra2 <- lapply(ymax, function(x) simulate_spec(wl_peak = 500, ylim = x))  # Simulate
spectra2 <- Reduce(function(...) merge(..., all = TRUE), spectra2)  # Combine
plot(spectra2)  # Plot

}
\author{
Thomas White \email{thomas.white026@gmail.com}

Hugo Gruson \email{hugo.gruson+R@normalesup.org}
}
