% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_spec.R
\name{simulate_spec}
\alias{simulate_spec}
\title{Simulate spectra}
\usage{
simulate_spec(
  shape = c("sigmoid", "bell"),
  hue,
  brightness,
  saturation,
  baseline,
  lim = c(300, 700)
)
}
\arguments{
\item{shape}{Simulate either a \code{"sigmoid"} ("yellow", "orange", "red", and
"white" colours) or a \code{"bell"}-shaped curve ("green", "blue", and "violet"
colours)}

\item{hue, brightness, saturation}{Colourimetric variable of the desired output}

\item{baseline}{Constant background reflectance for all wavelengths}

\item{lim}{A vector with the range of the wavelengths on which to evaluate
the function. Defaults to \code{c(300, 700)}}
}
\value{
A \code{rspec} object
}
\description{
Simulate spectra
}
\details{
Spectra are simulated with the following functions:
\itemize{
\item \code{"sigmoid"}:
\deqn{R(\text{wl})=\text{baseline}+\text{brightness} \times\exp \left[-\dfrac{(\text{wl}-\text{hue})^2}{2\text{saturation}^{2}}\right]}{R(wl) = baseline + brightness * exp(-(wl-hue)^2/(2*saturation^2))}
\item \code{"bell"}:
\deqn{R(\text{wl})=\text{baseline}+\text{brightness}\left[\frac{\pi}{2}+\arctan \left(\frac{\text{wl}-\text{hue}}{\text{saturation}}\right)\right]}{R(wl) = baseline + brightness * (pi/2 + arctan((wl-hue)/sigma))}
}
}
\examples{
plot(simulate_spec("sigmoid", 450, 35, 10, 5))
plot(simulate_spec("bell", 450, 35, 10, 5))

}
\references{
Gomez D., & Théry M. (2007), Simultaneous Crypsis and
Conspicuousness in Color Patterns: Comparative Analysis of a Neotropical
Rainforest Bird Community. The American Naturalist 169(S1), S42-S61.
\doi{10.1086/510138}
}
