% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rimg.R
\name{summary.rimg}
\alias{summary.rimg}
\title{Colour-classified image summary}
\usage{
\method{summary}{rimg}(object, plot = FALSE, ...)
}
\arguments{
\item{object}{(required) Results of \code{\link{classify}}.}

\item{plot}{logical; plot both the image and the colours corresponding to colour class
categories side-by-side? Defaults to \code{FALSE}.}

\item{...}{additional graphical options when \code{plot = TRUE}. Also see \code{\link{par}}.}
}
\value{
Either the RGB values of the k-means centres from the colour-classified image,
or a plot of both the image and specified colours (when \code{plot = TRUE}.
}
\description{
Returns the attributes of the colour-classified matrix generated from \code{\link{classify}}.
}
\examples{
\dontrun{
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))
papilio_class <- classify(papilio, kcols = 4)
summary(papilio_class)

# Plot the colour-classified image alongside the colour class palette
summary(papilio_class, plot = TRUE)
}

}
\author{
Thomas E. White \email{thomas.white026@gmail.com}
}
