% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcspoints.R
\name{.tcspoints}
\alias{.tcspoints}
\title{Plot a Tetrahedral Color Space}
\usage{
.tcspoints(tcsdata, size = 0.02, col = "black", alpha = 1)
}
\value{
\code{tcspoints} creates 3D points in a tetrahedral color space plot produced by \code{tcsplot}
using functions of the package \code{rgl}, based on openGL capabilities.
}
\description{
\code{tcspoints} plots points in a tetrahedral color space
}

