% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcspoints.R, R/tcsvol.R
\name{tcspoints}
\alias{tcspoints}
\alias{tcsvol}
\title{Plot a Tetrahedral Color Space}
\usage{
tcspoints(tcsdata, size = 0.02, col = "black", alpha = 1)

tcsvol(tcsdata, col = "black", alpha = 0.2, grid.alpha = 1, grid = T,
  fill = T)
}
\value{
\code{tcspoints} creates 3D points in a tetrahedral color space plot produced by \code{tcsplot}
using functions of the package \code{rgl}, based on openGL capabilities.

\code{tcsvol} creates a 3D convex hull within a \code{tcsplot} object
}
\description{
\code{tcspoints} plots points in a tetrahedral color space

\code{tcsvol} produces a 3D convex hull in tetrahedral color space
}

