% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-vissyst.R
\docType{data}
\name{vissyst}
\alias{vissyst}
\title{Animal visual systems data}
\description{
Internal data for visual model calculations.
}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}
}
\references{
Endler, J. A., & Mielke, P. (2005). Comparing entire colour patterns as birds see them. Biological Journal Of The Linnean Society, 86(4), 405-431.
}
\keyword{data}

