% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol.R
\name{vol}
\alias{vol}
\title{Plot a Tetrahedral Color Space}
\usage{
vol(tcsdata, alpha = 0.2, grid = TRUE, fill = TRUE, xlim = NULL,
  ylim = NULL, zlim = NULL, view = NULL, scale.y = NULL,
  margin = NULL, axis = FALSE, new = FALSE, ...)
}
\arguments{
\item{tcsdata}{(required) object of class \code{colspace}.}

\item{alpha}{transparency of volume (if \code{fill = TRUE}).}

\item{grid}{logical. if \code{TRUE} (default), draws the polygon outline defined by the points.}

\item{fill}{logical. if \code{TRUE} (default), fills the volume defined by the points.}

\item{xlim, ylim, zlim, view, scale.y, margin, axis}{plotting parameters 
in case of a new plot. see \code{\link{tetraplot}}.}

\item{new}{logical. Should a new plot be started or draw over an open plot? 
(defaults to FALSE)}

\item{...}{additional graphical options. See \code{link{polygon}}.}
}
\value{
\code{vol} creates a 3D convex hull within a static tetrahedral plot
}
\description{
\code{vol} produces a 3D convex hull in tetrahedral color space when plotting a 
non-interactive tetrahedral plot
}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}
}

