% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol.R
\name{vol}
\alias{vol}
\title{Plot a Tetrahedral Color Space}
\usage{
vol(colspacedata, alpha = 0.2, grid = TRUE, fill = TRUE, ...)
}
\arguments{
\item{colspacedata}{(required) and object of class \code{colspace}.}

\item{alpha}{transparency of volume (if \code{fill = TRUE}).}

\item{grid}{logical. if \code{TRUE} (default), draws the polygon outline defined by the points.}

\item{fill}{logical. if \code{TRUE} (default), fills the volume defined by the points.}

\item{...}{additional graphical options. See \code{link{polygon}}.}
}
\value{
\code{vol} creates a 3D convex hull within a static tetrahedral plot
}
\description{
\code{vol} produces a 3D convex hull in tetrahedral color space when plotting a 
non-interactive tetrahedral plot
}
\author{
Rafael Maia \email{rm72@zips.uakron.edu}
}

