% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol.R
\name{vol}
\alias{vol}
\title{Plot a 3D Convex Hull in a Tetrahedral Colorspace}
\usage{
vol(x, ...)
}
\arguments{
\item{...}{additional graphical options. See \code{\link{par}}.}

\item{clrspdata}{(required) an object of class \code{colspace}.}
}
\value{
\code{vol.colspace} creates a 3D convex hull within a \code{plot.colspace} object, 
 only when \strong{\code{space = 'tcs'}}.
}
\description{
\code{vol} produces a 3D convex hull in tetrahedral color space (i.e. when 
\code{space = 'tcs'})
}

